refactor(components): [steps] switch to script-setup syntax (#7783)
* refactor(components): [steps] switch to script-setup syntax

improvement(components): [steps] props type

improvement(components): [steps] unnecessary type

* fix(comoponents): [steps] import error

* fix(comoponents): [steps] import error
