+++ packages/components/steps/index.ts
@@ -1,6 +1,6 @@
 import { withInstall, withNoopInstall } from '@element-plus/utils'
 
-import Steps from './src/index.vue'
+import Steps from './src/steps.vue'
 import Step from './src/item.vue'
 
 export const ElSteps = withInstall(Steps, {
